r.profile: fixed failing tests on 64bit - using double instead of float, add additional test
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@62299 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.profile/main.c
@@ -272,7 +272,7 @@ int do_profile(double e1, double e2, double n1, double n2,
 	       int coords, double res, int fd, int data_type, FILE * fp,
 	       char *null_string, const char *unit, double factor)
 {
-    float rows, cols, LEN;
+    double rows, cols, LEN;
     double Y, X, AZI;
 
     cols = e1 - e2;
